Refactor and rename LogsDeliveryPath -> DirectoryPath
