+++ config/crd/bases/cluster.ytsaurus.tech_ytsaurus.yaml
@@ -19687,9 +19687,9 @@ spec:
                     type: integer
                   timbertruck:
                     properties:
-                      image:
+                      directoryPath:
                         type: string
-                      logsDeliveryPath:
+                      image:
                         type: string
                     type: object
                   tolerations:
@@ -32643,9 +32643,9 @@ spec:
                       type: integer
                     timbertruck:
                       properties:
-                        image:
+                        directoryPath:
                           type: string
-                        logsDeliveryPath:
+                        image:
                           type: string
                       type: object
                     tolerations:
